web/requirements.txt
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-Session==0.3.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
urllib3==1.25.10
Werkzeug==1.0.1
